Switch to a Gradle build with a two-subprojects source layout
Now the source repository has two subprojects, a 'echoxsl' and a 'serializer'.

In a future commit, the names of the directories shall be changed to module names.

The package names have not been switched to EchoXSL names yet.

The tests were added as a subtree of the 'echoxsl' subproject. The tests are being
compiled but not executed (a few test files had to be deleted to achieve that).